change primary keys to bigint

+++ app/schemas.php
@@ -6,7 +6,7 @@ $primarySchema = new Schema();
 
 $cardTable = $primarySchema->createTable('cards');
 
-$cardTable->addColumn('id', 'integer', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
+$cardTable->addColumn('id', 'bigint', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
 $cardTable->setPrimaryKey(['id']);
 
 $cardTable->addColumn('name', 'string', ['length' => 255, 'notnull' => true]);
@@ -18,7 +18,7 @@ $cardTable->addColumn('updated_at', 'datetime', ['notnull' => true]);
 
 $scheduleTable = $primarySchema->createTable('schedules');
 
-$scheduleTable->addColumn('id', 'integer', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
+$scheduleTable->addColumn('id', 'bigint', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
 $scheduleTable->setPrimaryKey(['id']);
 
 $scheduleTable->addColumn('name', 'string', ['length' => 255, 'notnull' => true]);
@@ -36,14 +36,14 @@ $scheduleTable->addColumn('updated_at', 'datetime', ['notnull' => true]);
 
 $cardScheduleTable = $primarySchema->createTable('card_schedule');
 
-$cardScheduleTable->addColumn('card_id', 'integer', ['length' => 20, 'unsigned' => true, 'notnull' => true]);
-$cardScheduleTable->addColumn('schedule_id', 'integer', ['length' => 20, 'unsigned' => true, 'notnull' => true]);
+$cardScheduleTable->addColumn('card_id', 'bigint', ['length' => 20, 'unsigned' => true, 'notnull' => true]);
+$cardScheduleTable->addColumn('schedule_id', 'bigint', ['length' => 20, 'unsigned' => true, 'notnull' => true]);
 
 $logSchema = new Schema();
 
 $logTable = $logSchema->createTable('logs');
 
-$logTable->addColumn('id', 'integer', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
+$logTable->addColumn('id', 'bigint', ['unsigned' => true, 'notnull' => true, 'autoincrement' => true, 'length' => 20]);
 $logTable->setPrimaryKey(['id']);
 
 $logTable->addColumn('code', 'string', ['length' => 6, 'notnull' => true]);
